generate_3d_model-main/requirements.txt
torch
torchvision
torchaudio
rembg
matplotlib
pyrender
open3d
ipywidgets
Pillow
opencv-python
trimesh
timm
pyrender
matplotlib
git+https://github.com/openai/shap-e.git
